Fix an accidental float truncation
This was causing some curves that needed barely more than N segments to
be rendered with N instead of N+1.

Reviewed-on: https://skia-review.googlesource.com/c/skia/+/467596
Commit-Queue: Chris Dalton <csmartdalton@google.com>
Commit-Queue: Greg Daniel <egdaniel@google.com>
Auto-Submit: Chris Dalton <csmartdalton@google.com>

+++ src/gpu/tessellate/PathCurveTessellator.cpp
@@ -101,7 +101,7 @@ public:
         }
     }
 
-    int numFixedSegments_pow4() const { return fNumFixedSegments_pow4; }
+    float numFixedSegments_pow4() const { return fNumFixedSegments_pow4; }
 
 private:
     void chopAndWriteQuadratic(const GrShaderCaps& shaderCaps, GrVertexChunkBuilder* chunker,
