+++ src/gpu/tessellate/PathWedgeTessellator.cpp
@@ -283,7 +283,7 @@ public:
         }
     }
 
-    int numFixedSegments_pow4() const { return fNumFixedSegments_pow4; }
+    float numFixedSegments_pow4() const { return fNumFixedSegments_pow4; }
 
 private:
     template <typename T>
